+++ lib_cicd_github/__init__conf__.py
@@ -2,7 +2,7 @@
 
 name = 'lib_cicd_github'
 title = 'CI/CD (Continuous Integration / Continuous Delivery) - utils for github actions'
-version = 'v1.0.2'
+version = 'v1.0.3'
 url = 'https://github.com/bitranox/lib_cicd_github'
 author = 'Robert Nowotny'
 author_email = 'bitranox@gmail.com'
@@ -16,7 +16,7 @@ Info for lib_cicd_github:
 
     CI/CD (Continuous Integration / Continuous Delivery) - utils for github actions
 
-    Version : v1.0.2
+    Version : v1.0.3
     Url     : https://github.com/bitranox/lib_cicd_github
     Author  : Robert Nowotny
     Email   : bitranox@gmail.com""")